Lights with Screen On [2/2]
Implement a setting allowing lights to be activated
for new notifications even if the screen is on.

Lights with screen on and Custom values are separated
in an advanced section for a cleaner overview.

This setting gives the user an oportunity to activate lights
with the screen on and also during DayDream screensaver.
The option is not activated by default.

Change-Id: I3650a014554ce50eb9030cc41e5ce1780bcc2dfe
Signed-off-by: AdrianDC <radian.dc@gmail.com>

Conflicts:
	res/values/du_strings.xml
	src/com/android/settings/notificationlight/NotificationLightSettings.java